backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
resend==2.0.0
